Retry resizing until quality is 15
The resizer should attempt to fit the resize into the desired size.
One way to do this, if the first attempt does not succeed, is to keep
trying by lowering the quality until we hit a pre-defined floor of 15.
Return the same error as before if even that doesn't work.